#139 data synchronisation reworked [x] server returns final change data for entities or if data should be pulled again [x] app waits until synchronization is done, before data is displayed again. This avoids merge problems

+++ sormas-app/app/src/main/java/de/symeda/sormas/app/backend/epidata/EpiDataDao.java
@@ -46,7 +46,9 @@ public class EpiDataDao extends AbstractAdoDao<EpiData> {
     @Override
     public boolean save(EpiData epiData) {
         try {
-            super.save(epiData);
+            if (!super.save(epiData)) {
+                return false;
+            }
             DatabaseHelper.getEpiDataBurialDao().deleteOrphansOfEpiData(epiData);
             if (epiData.getBurials() != null && !epiData.getBurials().isEmpty()) {
                 for (EpiDataBurial burial : epiData.getBurials()) {
